backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# LangChain + Gemini
langchain==0.2.12
langchain-google-genai==1.0.7
langchain-community==0.2.11

# Google AI
google-generativeai==0.7.2

# Tools
httpx==0.27.0           # async HTTP for API calls
duckduckgo-search==6.2.6  # free web search
pypdf==4.3.1            # document summarization
python-docx==1.1.2

# Code execution sandbox
RestrictedPython==7.0

# Utilities
pydantic==2.8.2
pydantic-settings==2.3.4
python-dotenv==1.0.1
